requirements.txt
boto3>=1.26.0
pyiceberg>=0.4.0
pandas>=2.0.0
pyarrow>=12.0.0
python-dotenv>=1.0.0
numpy>=1.24.0 